feat: 区分sync命令
